requirements.txt
fastapi
uvicorn
motor
pymongo
python-dotenv
pydantic

groq


pymupdf
python-docx

python-multipart
fastapi-mail

PyJWT